+++ pkg/querier/queryrange/downstreamer.go
@@ -262,6 +262,15 @@ func ResponseToResult(resp queryrangebase.Response) (logqlmodel.Result, error) {
 			Data:    matrix,
 			Headers: resp.GetHeaders(),
 		}, nil
+	case *QuantileSketchResponse:
+		matrix, err := sketch.QuantileSketchMatrixFromProto(r.Response)
+		if err != nil {
+			return logqlmodel.Result{}, fmt.Errorf("cannot decode quantile sketch: %w", err)
+		}
+		return logqlmodel.Result{
+			Data:    matrix,
+			Headers: resp.GetHeaders(),
+		}, nil
 	default:
 		return logqlmodel.Result{}, fmt.Errorf("cannot decode (%T)", resp)
 	}