+++ pkg/storage/config/store.go
@@ -15,7 +15,7 @@ import (
 type ChunkStoreConfig struct {
 	ChunkCacheConfig       cache.Config `yaml:"chunk_cache_config"`
 	ChunkCacheConfigL2     cache.Config `yaml:"chunk_cache_config_l2" doc:"hidden"`
-	WriteDedupeCacheConfig cache.Config `yaml:"write_dedupe_cache_config"`
+	WriteDedupeCacheConfig cache.Config `yaml:"write_dedupe_cache_config" doc:"description=Write dedupe cache is deprecated along with legacy index types (aws, aws-dynamo, bigtable, bigtable-hashed, cassandra, gcp, gcp-columnkey, grpc-store).\nConsider using TSDB index which does not require a write dedupe cache."`
 
 	L2ChunkCacheHandoff   time.Duration  `yaml:"l2_chunk_cache_handoff" doc:"hidden"`
 	CacheLookupsOlderThan model.Duration `yaml:"cache_lookups_older_than"`
@@ -55,8 +55,6 @@ func (cfg *ChunkStoreConfig) Validate(logger log.Logger) error {
 		flagext.DeprecatedFlagsUsed.Inc()
 		level.Warn(logger).Log("msg", "running with DEPRECATED flag -store.max-look-back-period, use -querier.max-query-lookback instead.")
 	}
-	if err := cfg.ChunkCacheConfig.Validate(); err != nil {
-		return err
-	}
-	return cfg.WriteDedupeCacheConfig.Validate()
+
+	return nil
 }